Move the CustomerRepository to the CustomerBundle
