+++ src/Sylius/Behat/Context/Ui/Shop/RegistrationContext.php
@@ -27,7 +27,7 @@ use Sylius\Behat\Service\NotificationCheckerInterface;
 use Sylius\Behat\Service\SharedStorageInterface;
 use Sylius\Component\Core\Model\CustomerInterface;
 use Sylius\Component\Core\Model\ShopUserInterface;
-use Sylius\Component\Core\Repository\CustomerRepositoryInterface;
+use Sylius\Component\Customer\Repository\CustomerRepositoryInterface;
 use Webmozart\Assert\Assert;
 
 class RegistrationContext implements Context
