+++ src/Sylius/Bundle/CoreBundle/Doctrine/ORM/CustomerRepository.php
@@ -13,46 +13,15 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\Doctrine\ORM;
 
-use Sylius\Bundle\ResourceBundle\Doctrine\ORM\EntityRepository;
+use Sylius\Bundle\CustomerBundle\Doctrine\ORM\CustomerRepository as BaseCustomerRepository;
 use Sylius\Component\Core\Model\CustomerInterface;
-use Sylius\Component\Core\Repository\CustomerRepositoryInterface;
 
 /**
  * @template T of CustomerInterface
  *
- * @implements CustomerRepositoryInterface<T>
+ * @extends BaseCustomerRepository<T>
  */
-class CustomerRepository extends EntityRepository implements CustomerRepositoryInterface
+class CustomerRepository extends BaseCustomerRepository
 {
-    public function countCustomers(): int
-    {
-        return (int) $this->createQueryBuilder('o')
-            ->select('COUNT(o.id)')
-            ->getQuery()
-            ->getSingleScalarResult()
-        ;
-    }
 
-    public function countCustomersInPeriod(\DateTimeInterface $startDate, \DateTimeInterface $endDate): int
-    {
-        return (int) $this->createQueryBuilder('o')
-            ->select('COUNT(o.id)')
-            ->where('o.createdAt >= :startDate')
-            ->andWhere('o.createdAt <= :endDate')
-            ->setParameter('startDate', $startDate)
-            ->setParameter('endDate', $endDate)
-            ->getQuery()
-            ->getSingleScalarResult()
-        ;
-    }
-
-    public function findLatest(int $count): array
-    {
-        return $this->createQueryBuilder('o')
-            ->addOrderBy('o.createdAt', 'DESC')
-            ->setMaxResults($count)
-            ->getQuery()
-            ->getResult()
-        ;
-    }
 }
