+++ src/Sylius/Bundle/CustomerBundle/Doctrine/ORM/CustomerRepository.php
@@ -1,0 +1,58 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Sylius Sp. z o.o.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\Bundle\CustomerBundle\Doctrine\ORM;
+
+use Sylius\Bundle\ResourceBundle\Doctrine\ORM\EntityRepository;
+use Sylius\Component\Customer\Model\CustomerInterface;
+use Sylius\Component\Customer\Repository\CustomerRepositoryInterface;
+
+/**
+ * @template T of CustomerInterface
+ *
+ * @implements CustomerRepositoryInterface<T>
+ */
+class CustomerRepository extends EntityRepository implements CustomerRepositoryInterface
+{
+    public function countCustomers(): int
+    {
+        return (int) $this->createQueryBuilder('o')
+            ->select('COUNT(o.id)')
+            ->getQuery()
+            ->getSingleScalarResult()
+        ;
+    }
+
+    public function countCustomersInPeriod(\DateTimeInterface $startDate, \DateTimeInterface $endDate): int
+    {
+        return (int) $this->createQueryBuilder('o')
+            ->select('COUNT(o.id)')
+            ->where('o.createdAt >= :startDate')
+            ->andWhere('o.createdAt <= :endDate')
+            ->setParameter('startDate', $startDate)
+            ->setParameter('endDate', $endDate)
+            ->getQuery()
+            ->getSingleScalarResult()
+        ;
+    }
+
+    public function findLatest(int $count): array
+    {
+        return $this->createQueryBuilder('o')
+            ->addOrderBy('o.createdAt', 'DESC')
+            ->setMaxResults($count)
+            ->getQuery()
+            ->getResult()
+        ;
+    }
+}
