+++ src/Sylius/Component/Customer/Repository/CustomerRepositoryInterface.php
@@ -11,9 +11,9 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Core\Repository;
+namespace Sylius\Component\Customer\Repository;
 
-use Sylius\Component\Core\Model\CustomerInterface;
+use Sylius\Component\Customer\Model\CustomerInterface;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
 
 /**
